Refactor Background
